requirements.txt
transformers>=4.35.0
accelerate>=0.25.0
pillow>=10.0.0
runpod>=1.6.0
safetensors>=0.4.0
git+https://github.com/huggingface/diffusers.git
